refactor(apps/client): HASHI-165 auth session을 shared로 이동
